+++ odoo/songs/install/data_full.py
@@ -44,6 +44,24 @@ def import_suppliers(ctx):
 
 
 @anthem.log
+def import_crm_team(ctx):
+    """ Importing sales team / channel from csv """
+    load_csv(ctx, 'data/demo/crm_team.csv', 'crm.team')
+
+
+@anthem.log
+def import_location(ctx):
+    """ Importing stock location from csv """
+    load_csv(ctx, 'data/demo/stock_location.csv', 'stock.location')
+
+
+@anthem.log
+def import_drone_type(ctx):
+    load_csv(ctx, 'data/install/drone_type.csv', 'drone.type')
+
+
+
+@anthem.log
 def import_product(ctx):
     load_csv(ctx, 'data/install/product.csv', 'product.template')
 
@@ -56,5 +74,8 @@ def main(ctx):
     import_countries(ctx)
     import_customers(ctx)
     import_suppliers(ctx)
+    import_crm_team(ctx)
+    import_location(ctx)
+    import_drone_type(ctx)
     import_product(ctx)
     return